Add reference support for designs
Adds necessary scopes (search and pre-fetching) to designs, and adds
tests for scopes and design referencing.

Design referencing is made rather more complex by the need to safely
escape the wide domain of filenames, which may contain any arbitrary
text values, and could even pose security issues if unescaped.

We deal with the need to handle embedded quotes and angle brackets with
quoting and escaping, and with more complex cases (such as HTML
entities) by escaping to base-64.

The scheme adopted here allows multiple different encodings to be used,
and all to be matched against at the same time.