Improve abstract reference filter
* Allow find_parent to be overridden in child classes

This allows the mechanism for resolving a parent to be overridden in a
child class, in case a more efficient mechanism exists.

* Allow reference filters to override symbol parsing

This involves safe extraction from match data. We expect this to be
consistent throughout the different lazily-constructed indexes, so we
override parsing in the commit filter, which does not need the default
int-ification.

* Improve abstract_reference_filter spec

This ensures we don't call object_sym any more than twice (hard-coded
limit in test that we don't need to exceed), and we add a test to check
that the caching works as expected.